requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.6
colorama==0.4.4
Django==3.0
django-auto-prefetching==0.1.10
django-colorfield==0.4.3
django-environ==0.7.0
django-filter==2.4.0
django-guardian==2.4.0
django-map-widgets==0.3.2
django-silk==4.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
drf-base64==2.0
flake8==4.0.1
gprof2dot==2021.2.21
gunicorn==20.0.4
idna==3.2
importlib-metadata==4.8.1
iniconfig==1.1.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.0
Pillow==8.3.2
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.10.0
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.6
zipp==3.5.0
